FT81X_LIB PLATFORM K66: Added a display spi edma handle pointer.
Currently we have the display and the gpu on the same SPI bus.
Meaning we don't need to set up seperate DMA handles for it.
It is possible to use separate SPI buses, in which case we would need
different handles.

In the case with the same SPI module for both, we can just set the
display handle to be a pointer to the gpu one.

In the case with different SPI modules I've added a #error TODO.