FT81X_LIB: Added a ft81x_send_display_list() function.
This allows the user to send a display list to the screen.

However there's a bug somewhere ATM which means only the first display
list actually works. Which is the init_dl sent in the configuration
stage.

Pretty sure this bug is not related to the ft81x_send_display_list()
function, but something to do with the GPU configuration.