FT81X_LIB: Added the ability to send a display list in chunks rather than all at once.
If you have a large display list that uses the image manager code to
assign a bitmap ID. You need to get that bitmap id first before you can
load an image in the DL or draw it.
So you can't just have a static display list defined as a global const
array. Since you don't have the bitmap IDs at that time. Meaning it all
has to be on the stack.
To avoid this we can now pass the display list in bit by bit. So static
sectons can be in global const arrays, and the rest can be split into
small functions so that there is never too much data on the stack.