FT81X_LIB Image Manager: Removed some more #warnings TODOs.
These warnings suggested that we use CMD_INTERRUPT to get an interrutp
after the inflate was done, so we could detect overflows without
blocking.

Unfortunately due to the issue discussed in the previous commit. You
can't write CMD_GETPTR until the inflate has finished. Meaning we would
have to write it in the interrupt. Any other inflate commands (or
possibly memory writes or even accesses) could break this functionality.
So you'd have to disable writing a bunch of commands until after the
interrupt, which would most likely mean blocking for ages anyway.