FT81X_LIB: Changed VOFFSET and HOFFSET register values.
I had an issue where only my first DL would display, and REG_FRAMES was
stuck at 0.
I opened a support ticket with FTDI and got the answer that:
VCYCLE must be greater than (VOFFSET + VSIZE).
According to all the documentation VCYCLE should be equal to VOFFSET +
VSIZE. However that is what caused my issue.

It is still unclear why this is the case, but it is a requirement.
I fix this simply by setting VCYCLE to the total lines + 1. Which works
perfectly.

The same is true of HCYCLE according to support, although I experienced
no issues with that. I added a +1 there as well to prevent any
unexpected issues.